+++ test_suite/bindir.py
@@ -1,0 +1,1 @@
+BINDIR = '/home/mkumar/process_reinke/bin'